gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
redis[hiredis]==5.0.7
qdrant-client==1.9.1
neo4j==5.21.0
boto3==1.34.131
aiobotocore==2.13.0
clickhouse-connect==0.7.16
aiokafka==0.11.0
PyYAML==6.0.1
prometheus-client==0.20.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
python-multipart==0.0.9
httpx==0.27.0
